requirements.txt
django>=2.2.10
Pillow==6.2.1
pytz==2018.9

# tools
django-extensions==2.2.8
django-cors-headers
# Environment
django-environ==0.4.5

# Django REST framework
djangorestframework==3.11.0

# JWT
pyjwt==1.7.1

# Calidad del código
flake8==3.6.0

# Decoplamiento Virtual
python-decouple
# Este es conector con postgress
# psycopg2==2.7.4 --no-binary psycopg2